requirements.txt
asgiref==3.4.1
certifi==2020.6.20
click==8.0.3
cmake==3.22.1
colorama==0.4.4
Django==3.2.10
django-crispy-forms==1.13.0
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==4.10.1
numpy==1.21.5
opencv-python==4.5.5.62
Pillow==8.4.0
psycopg2==2.9.2
pytz==2020.1
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.7.0

